PerAlleleAnnotation is Integer, not Float

+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/annotator/PerAlleleAnnotation.java
@@ -66,7 +66,7 @@ public abstract class PerAlleleAnnotation extends GenotypeAnnotation {
 
     @Override
     public List<VCFFormatHeaderLine> getDescriptions() {
-        return Arrays.asList(new VCFFormatHeaderLine(getVcfKey(), includeRefAllele() ? VCFHeaderLineCount.R : VCFHeaderLineCount.A, VCFHeaderLineType.Float, getDescription()));
+        return Arrays.asList(new VCFFormatHeaderLine(getVcfKey(), includeRefAllele() ? VCFHeaderLineCount.R : VCFHeaderLineCount.A, VCFHeaderLineType.Integer, getDescription()));
     }
 
     @Override
